commit aefa9c84b0a900bedba3a7ed885f0ea75f3fe1ea
    12015 vioif with MSI-X not working on Google Compute Engine
commit c00b0288a75782c75abd0b9959021fd2417c8a28
    12014 virtio_init_complete() comment should describe interrupt type parameter
commit d1f3e3c647f14e2496e646bb7ef0df729cbbf72a
    11798 apix: cast between incompatible function types
commit 027bcc9f64a0a5915089267b0dc54c9ee05782b0
    11797 i86pc: cast between incompatible function types

+++ usr/src/uts/intel/ia32/os/desctbls.c
@@ -103,7 +103,7 @@ user_desc_t	*gdt0;
 desctbr_t	gdt0_default_r;
 #endif
 
-gate_desc_t	*idt0; 		/* interrupt descriptor table */
+gate_desc_t	*idt0;		/* interrupt descriptor table */
 #if defined(__i386)
 desctbr_t	idt0_default_r;		/* describes idt0 in IDTR format */
 #endif
@@ -147,10 +147,10 @@ void (*(fasttable[]))(void) = {
 	fast_null,			/* T_FNULL routine */
 	fast_null,			/* T_FGETFP routine (initially null) */
 	fast_null,			/* T_FSETFP routine (initially null) */
-	(void (*)())get_hrtime,		/* T_GETHRTIME */
-	(void (*)())gethrvtime,		/* T_GETHRVTIME */
-	(void (*)())get_hrestime,	/* T_GETHRESTIME */
-	(void (*)())getlgrp		/* T_GETLGRP */
+	(void (*)())(uintptr_t)get_hrtime,	/* T_GETHRTIME */
+	(void (*)())(uintptr_t)gethrvtime,	/* T_GETHRVTIME */
+	(void (*)())(uintptr_t)get_hrestime,	/* T_GETHRESTIME */
+	(void (*)())(uintptr_t)getlgrp		/* T_GETLGRP */
 };
 
 /*
@@ -1356,7 +1356,7 @@ void
 brand_interpositioning_enable(void)
 {
 	gate_desc_t	*idt = CPU->cpu_idt;
-	int 		i;
+	int		i;
 
 	ASSERT(curthread->t_preempt != 0 || getpil() >= DISP_LEVEL);
